requirements.txt
flask==2.3.3
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.25.0
regex==2023.6.3
openai==1.12.0
faiss-cpu==1.7.4
